feat: add `verifyContract` function
