+++ axelar-chains-config/src/utils/index.js
@@ -1,3 +1,4 @@
 export * from './getChainArray';
 export * from './readJSON';
 export * from './importNetworks';
+export * from './verifyContract';
